Implemented features: task that updates the most needy prices of each n sellers every five minutes or so. -- IMPORTANT: need a new properties file.
New properties file is called:

/WEB-INF/local.properties

Should look like this:
debug.enabled=true
pricetask.enabled=false

So it doesn't update prices locally all the time.

On another note, there was a problem with selecting certain columns in the view so I had to switch away from the view and just create the SELECT query on-the-fly. Should definitely be changed later.